udebug-remote: pass size_t to calloc_a()
udebug_remote_buf_snapshot() passed data_size, a uint32_t, as a chunk
size to calloc_a(). __calloc_a() reads chunk sizes as size_t through
varargs, so on LP64 systems va_arg(ap, size_t) reads 8 bytes where
only 4 were pushed. Cast data_size to size_t at the call site.

Link: https://github.com/openwrt/libubox/pull/46
(cherry picked from commit 9afc710534817a3066e42542b3ef07ccd1e903a7)

+++ udebug-remote.c
@@ -237,7 +237,7 @@ udebug_remote_buf_snapshot(struct udebug_remote_buf *rb)
 
 	s = calloc_a(sizeof(*s),
 		     &ptr_buf, ptr_size * sizeof(*ptr_buf),
-		     &data_buf, data_size);
+		     &data_buf, (size_t)data_size);
 
 	s->data = memcpy(data_buf, udebug_buf_ptr(&rb->buf, data_start), data_size);
 	s->data_size = data_size;
